Allow specifying buffer subsets.
Sometimes audiobuffers contain data a user might not be interested to
measure.

Obviating the CPU waste of doing two different passes, one for the
magnitude and another for the RMS; touching memory areas that are not
required is not good for caches, specially if some channel buffers there
are only preallocated but unused.

This commit adds an overload to "measureBlock" that allows specifying
buffer subsets.